Fixing revert and creating 'happy path' unit tests. (#293)

+++ modules/custom/dkan_harvest/src/Harvester.php
@@ -100,9 +100,14 @@ class Harvester {
 
   /**
    * Public.
+   *
+   * @todo the destroy method should be part of some interface.
    */
   public function revertHarvest($id) {
     $run_store = $this->storeFactory->getInstance("harvest_{$id}_runs");
+    if (!method_exists($run_store, "destroy")) {
+      throw new \Exception("Storage of class " . get_class($run_store) . " does not implement destroy method.");
+    }
     $run_store->destroy();
     $harvester = $this->getHarvester($id);
     return $harvester->revert();
